Fix auto farm tests to use refactored farming exports
